Functioning step.
Code actions work.
published diagnostics are not rendered. The hx log shows them being
received and dropped.

Needed to replace the HashMap of LanguageServers with a Vector. The
orginal code looped over the LanguageConfigs for the document. It did
this to preserve the original ordering from the TOML according to the
comments. But with this branch's change there are more language servers
than the LanguageConfiguration knows about. This lead the aforementioned
loop to skip the __common__ servers. Opinion. The HashMap was not
buying much since there are likely to be less than 5 or 6 active
language servers per document. Often only 1 to 3. Looping and looking
in a Vec is not a big deal at this scale compared to generating hashes
and looking up entries.